+++ docs/datasets.rst
@@ -11,5 +11,5 @@ The hdc library provides many popular built-in datasets to work with.
 
     BeijingAirQuality
     ISOLET
-    LanguageRecognition
+    EuropeanLanguages
     UCIHAR
