+++ hdc/datasets/__init__.py
@@ -1,6 +1,6 @@
 from .beijing_air_quality import BeijingAirQuality
 from .isolet import ISOLET
-from .language_recognition import LanguageRecognition
+from .european_languages import EuropeanLanguages
 from .ucihar import UCIHAR
 
-__all__ = ["BeijingAirQuality", "ISOLET", "LanguageRecognition", "UCIHAR"]
+__all__ = ["BeijingAirQuality", "ISOLET", "EuropeanLanguages", "UCIHAR"]
